sm7325-common: properties: Set debug.sf.frame_rate_multiple_threshold to 60 by default
If the device only have 60 and 120 Hz in the config group, 60 Hz won't be
used when watching a 24 fps video and instead 120 Hz will be used because
it is perfectly divisible by 24.

This is a publicly documented AOSP behavior [1], but it simply does not
make any practical sense and trades off significant amount of power for a
very-hard-to-notice jitter-free experience.

Set debug.sf.frame_rate_multiple_threshold to 60 by default to limit this
behavior to 60 or lower fps.

[1] https://source.android.com/devices/graphics/multiple-refresh-rate

Change-Id: If88e877c6b74de457a5cf348fd376c77cf97cfc6
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>